refactor(dashboards): split records authoring helper
